fixed issue with flat signal in denoising step

+++ Functions/twoPhoton/get2pActivityMatrix.m
@@ -1,12 +1,13 @@
 function twopActivityMatrix = get2pActivityMatrix(F, denoisedF, spks, spikeTimes, resamplingRate, Info, Params)
 %GET2PACTIVITYMATRIX Summary of this function goes here
 %   Detailed explanation goes here
+% twopActivityMatrix (numTimeSmaples, numUnits)
 if strcmp(Params.twopActivity, 'F')
-    twopActivityMatrix = F;
+    twopActivityMatrix = double(F);
 elseif strcmp(Params.twopActivity, 'denoised F')
-    twopActivityMatrix = denoisedF;
+    twopActivityMatrix = double(denoisedF);
 elseif strcmp(Params.twopActivity, 'spks')
-    twopActivityMatrix = spks;
+    twopActivityMatrix = double(spks);
 elseif strcmp(Params.twopActivity, 'peaks')
     twopActivityMatrix = SpikeTimesToMatrix(spikeTimes, resamplingRate, 'peak', Info);
 end
